+++ js/modules/geom/TGeoPainter.mjs
@@ -1810,7 +1810,7 @@ class TGeoPainter extends ObjectPainter {
 
    /** @summary add transformation control */
    addTransformControl() {
-      if (this._tcontrols) return;
+      if (this._tcontrols || !this._webgl || isBatchMode()) return;
 
       if (!this.ctrl._debug && !this.ctrl._grid) return;