+++ js/modules/tree.mjs
@@ -1370,7 +1370,7 @@ class TDrawSelector extends TSelector {
             case 1:
                for (let n0 = 0; n0 < var0.length; ++n0) {
                   var0.buf.push(var0.get(n0));
-                  cut.buf.push(cut.value);
+                  cut.buf?.push(cut.value);
                }
                break;
             case 2:
@@ -1378,7 +1378,7 @@ class TDrawSelector extends TSelector {
                   for (let n1 = 0; n1 < var1.length; ++n1) {
                      var0.buf.push(var0.get(n0));
                      var1.buf.push(var1.get(n1));
-                     cut.buf.push(cut.value);
+                     cut.buf?.push(cut.value);
                   }
                break;
             case 3:
@@ -1388,7 +1388,7 @@ class TDrawSelector extends TSelector {
                         var0.buf.push(var0.get(n0));
                         var1.buf.push(var1.get(n1));
                         var2.buf.push(var2.get(n2));
-                        cut.buf.push(cut.value);
+                        cut.buf?.push(cut.value);
                      }
                break;
          }
@@ -1553,7 +1553,8 @@ async function treeProcess(tree, selector, args) {
 
    if (!selector || !tree.$file || !selector.numBranches()) {
       if (selector) selector.Terminate(false);
-      return Promise.reject(Error('required parameter missing for TTree::Process'));
+      console.error('required parameter missing for TTree::Process');
+      return null;
    }
 
    // central handle with all information required for reading
@@ -2141,7 +2142,8 @@ async function treeProcess(tree, selector, args) {
 
       if (!item) {
          selector.Terminate(false);
-         return Promise.reject(Error(`Fail to add branch ${selector.nameOfBranch(nn)}`));
+         console.error(`Fail to add branch ${selector.nameOfBranch(nn)}`);
+         return null;
       }
    }
 
@@ -2168,7 +2170,8 @@ async function treeProcess(tree, selector, args) {
 
    if (handle.firstentry >= handle.lastentry) {
       selector.Terminate(false);
-      return Promise.reject(Error('No any common events for selected branches'));
+      console.error('No any common events for selected branches');
+      return null;
    }
 
    handle.process_min = handle.firstentry;