+++ js/changes.md
@@ -2,6 +2,9 @@
 
 ## Changes in 7.3.x
 1. Fix - failure in normal_cdf calculation
+2. Fix - check in TTree::Draw for null buffer
+3. Fix - do not rise exception in treeProcess
+4. Fix - RH1 zero line drawing only when required
 
 
 ## Changes in 7.3.2